Update GetBufferSize() to return uint32_t (2^32) instead of size_t (2^64 on 64-bit). Add initial implementations of JsonFullWriter and JsonFullReader archivers.

+++ utils/BinaryArchiveReader.cpp
@@ -104,7 +104,7 @@ namespace Kernel
     bool BinaryArchiveReader::HasError() { return error; }
     bool BinaryArchiveReader::IsWriter() { return false; }
 
-    size_t BinaryArchiveReader::GetBufferSize()
+    uint32_t BinaryArchiveReader::GetBufferSize()
     {
         throw "BinaryArchiveReader doesn't implement GetBufferSize().";
     }
