+++ utils/JsonRawWriter.h
@@ -33,7 +33,7 @@ namespace Kernel
 
         virtual bool HasError() override;
         virtual bool IsWriter() override;
-        virtual size_t GetBufferSize() override;
+        virtual uint32_t GetBufferSize() override;
         virtual const char* GetBuffer() override;
 
         rapidjson::StringBuffer* m_buffer;
